refactor: add noexcept specifier
This commit adds noexcept specifier to methods which are known not to
fail or should not throw. This commit also replaces use legacy <math.h>
APIs with strengthened equivalents in <cmath>

One notable example is destructors which call Vulkan Memory Allocator
(VMA) APIs as part of RAII. Although VMA doesn't annotate their APIs
with noexcept, their public documentation explicitly mentions that "C++
exceptions are not used."

+++ src/graphics/camera.cpp
@@ -37,7 +37,7 @@ void Camera::Translate(const float dx, const float dy, const float dz) {
 void Camera::Rotate(const float theta, const float phi) {
   static constexpr auto kThetaMax = glm::two_pi<float>();
   static constexpr auto kPhiMax = glm::radians(89.0f);
-  orientation_.theta = std::fmodf(orientation_.theta + theta, kThetaMax);
+  orientation_.theta = std::fmod(orientation_.theta + theta, kThetaMax);
   orientation_.phi = std::clamp(orientation_.phi + phi, -kPhiMax, kPhiMax);
 }
